Reduce memory requirement for record
The identifiers for a csp and provider are generated by humans
developing give, and really only have meaning to us as a way to identify
csps and providers by name. Reducing their character limit from 60-16
characters still allows us to identify them with a meaningful name, that
doesn't take up as much extra space in the transaction database.