Remove superfluous view name parameter
When creating a new record, the view url will never change, so it
doesn't make sense to allow this as a parameter. The DRF router
registers URL patterns based off the TransactionRecord viewset, which
automatically makes the {basename}-list reverse() lookup available to
use as the endpoint for new Transaction objects.